♻️ create generic credential class
